grades_service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pymongo>=4.6.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
PyJWT>=2.8.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
